Fix sending server memory state at compilation failure

+++ runtime/compiler/control/JITServerCompilationThread.cpp
@@ -770,7 +770,7 @@ TR::CompilationInfoPerThreadRemote::processEntry(TR_MethodToBeCompiled &entry, J
       {
       if (TR::Options::getVerboseOption(TR_VerboseJITServer))
          TR_VerboseLog::writeLineLocked(TR_Vlog_JITServer, "Server out of memory in processEntry: %s", e.what());
-      stream->writeError(compilationLowPhysicalMemory, computeServerMemoryState(getCompilationInfo()));
+      stream->writeError(compilationLowPhysicalMemory, (uint64_t) computeServerMemoryState(getCompilationInfo()));
       abortCompilation = true;
       }
